Sube correcciones de controllers y sincronizacion con entidades

+++ src/main/java/com/kitchenhack/apikitchen/repositories/RecipeRepository.java
@@ -3,14 +3,12 @@ package com.kitchenhack.apikitchen.repositories;
 import com.kitchenhack.apikitchen.entities.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
-public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
     // Busca todas las recetas donde el campo 'publicada' sea true
     List<Recipe> findByPublishedTrue();
